# api/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Mahad Nafsy API",
    "version": "2.0.0",
    "description": "Auto-generated from api/routes/*.js by tools/gen-openapi.mjs. Operation bodies are stubs; enrich summaries/schemas as needed."
  },
  "servers": [
    {
      "url": "https://mahadnafsy.com"
    },
    {
      "url": "http://127.0.0.1:3001"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
...
    "/api/admin/orders/{id}": {
      "patch": {
        "tags": [
          "orders"
        ],
        "summary": "PATCH /api/admin/orders/:id",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "delete": {
        "tags": [
          "orders"
        ],
        "summary": "DELETE /api/admin/orders/:id",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    "/api/orders/reserve": {
      "post": {
        "tags": [
          "public-orders"
        ],
        "summary": "POST /api/orders/reserve",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    "/api/payments/paymob-init": {
      "post": {
        "tags": [
          "public-orders"
        ],
        "summary": "POST /api/payments/paymob-init",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/paymob/verify": {
      "post": {
        "tags": [
          "public-orders"
        ],
        "summary": "POST /api/paymob/verify",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    "/api/webhooks/paymob": {
      "post": {
        "tags": [
          "public-orders"
        ],
        "summary": "POST /api/webhooks/paymob",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    }
  }
}